+++ packages/xchain-evm/CHANGELOG.md
@@ -1,3 +1,10 @@
+# v0.4.0 (2023-11-15)
+
+## Update
+
+- Default gasPrice in baseAmount unit. Changed from GWei to Wei
+- Gas price is retrieved from provider as fallback if the getFeeRates round robin fails
+
 # v0.3.8 (2023-11-10)
 
 ## Update